(#106) Refactor PHP utilities and introduce color utilities * Introduce a `Color_Utils` class containing color utilities * Split the existing `Utils` class into three classes under a `Utils` sub-namespace:   * `String_Utils` - string utilities (e.g. truncate string)   * `Color_Utils` - color utilities (e.g. get hex color by a color slug)   * `WP_Utils` - WordPress utilities (e.g. get domain)
