+++ includes/Admin/Settings/Consent_Record.php
@@ -8,9 +8,9 @@
 
 namespace Pressidium\WP\CookieConsent\Admin\Settings;
 
-use Pressidium\WP\CookieConsent\Utils;
-
 use DateTime;
+
+use Pressidium\WP\CookieConsent\Utils\String_Utils;
 
 if ( ! defined( 'ABSPATH' ) ) {
     die( 'Forbidden' );
@@ -121,7 +121,7 @@ class Consent_Record {
      * @return Consent_Record
      */
     public function set_id( string $id ): Consent_Record {
-        $this->id = Utils::truncate( $id, 40 );
+        $this->id = String_Utils::truncate( $id, 40 );
 
         return $this; // chainable
     }
@@ -143,7 +143,7 @@ class Consent_Record {
      * @return Consent_Record
      */
     public function set_url( string $url ): Consent_Record {
-        $this->url = Utils::truncate( $url, 255 );
+        $this->url = String_Utils::truncate( $url, 255 );
 
         return $this; // chainable
     }
@@ -165,7 +165,7 @@ class Consent_Record {
      * @return Consent_Record
      */
     public function set_geo_location( ?string $geo_location ): Consent_Record {
-        $this->geo_location = Utils::truncate( $geo_location, 8 );
+        $this->geo_location = String_Utils::truncate( $geo_location, 8 );
 
         return $this; // chainable
     }
@@ -187,7 +187,7 @@ class Consent_Record {
      * @return Consent_Record
      */
     public function set_ip_address( string $ip_address ): Consent_Record {
-        $this->ip_address = Utils::truncate( wp_privacy_anonymize_ip( $ip_address ), 40 );
+        $this->ip_address = String_Utils::truncate( wp_privacy_anonymize_ip( $ip_address ), 40 );
 
         return $this; // chainable
     }
@@ -209,7 +209,7 @@ class Consent_Record {
      * @return Consent_Record
      */
     public function set_user_agent( string $user_agent ): Consent_Record {
-        $this->user_agent = Utils::truncate( $user_agent, 255 );
+        $this->user_agent = String_Utils::truncate( $user_agent, 255 );
 
         return $this; // chainable
     }
